Fix compilation errors on macOS

+++ sdlgui/imageview.cpp
@@ -300,7 +300,12 @@ void ImageView::draw(SDL_Renderer* renderer)
         positionAfterOffset.y = absolutePosition().y;
       }
       SDL_Rect imgrect{ix, iy, iw, ih};
-      SDL_Rect rect{ positionAfterOffset.x, positionAfterOffset.y, imgrect.w, imgrect.h};
+      SDL_Rect rect{
+        static_cast<int>(positionAfterOffset.x),
+        static_cast<int>(positionAfterOffset.y),
+        static_cast<int>(imgrect.w),
+        static_cast<int>(imgrect.h)
+      };
 
       SDL_RenderCopy(renderer, mTexture, &imgrect, &rect);
     }
@@ -368,7 +373,12 @@ void ImageView::drawHelpers(SDL_Renderer* renderer) const
   Vector2f sizeOffsetDifference = sizeF() - mOffset;
   Vector2f scissorSize = sizeOffsetDifference.cmin(sizeF());
 
-  SDL_Rect r{ scissorPosition.x, scissorPosition.y, scissorSize.x, scissorSize.y };
+  SDL_Rect r{
+    static_cast<int>(scissorPosition.x),
+    static_cast<int>(scissorPosition.y),
+    static_cast<int>(scissorSize.x),
+    static_cast<int>(scissorSize.y)
+  };
   if (gridVisible())
     drawPixelGrid(renderer, upperLeftCorner, lowerRightCorner, mScale);
   if (pixelInfoVisible())
